+++ src/compas/geometry/shapes/cone.py
@@ -340,6 +340,22 @@ class Cone(Shape):
         # Apply transformation to frame
         self.frame.transform(transformation)
 
+    def scale(self, factor):
+        """Scale the cone by multiplying the radius and height by a factor.
+
+        Parameters
+        ----------
+        factor : float
+            The scaling factor.
+
+        Returns
+        -------
+        None
+
+        """
+        self.radius *= factor
+        self.height *= factor
+
     # ==========================================================================
     # Methods
     # ==========================================================================